feat(util): introduce `channel::Sender::try_send()`
pr #140 introduced a new channel-backed body. this body provides a new
equivalent to the defunct `hyper::Body::channel()` interface that was
exposed in hyper 0.14, and removed in the 1.0 major release.

the previous `Sender` type also included a useful method,
`try_send_data()`, which allows a thread to _synchronously_ attempt to
send data _outside_ of an asynchronous context.

<https://docs.rs/hyper/0.14.31/hyper/body/struct.Sender.html#method.try_send_data>

this commit introduces a loosely equivalent `Sender::try_send()` method
to provide users of `http-body-util` with a means to send `Frame<T>`
values outside of an asynchronous context, without potentially
scheduling the caller to be woken later or yielding should the channel
be full.

this function accepts a `Frame<D>` rather than a chunk of data, to fit
in with the shift towards frame-oriented interfaces in `http-body` and
`http-body-util`.

Signed-off-by: katelyn martin <me+cratelyn@katelyn.world>